Updated files to use d4 instead of Knickers
Using the name "knickers," while cute, made it so that all files,
functions, and hooks had to be renamed before uploading to a client
site. Using standardized D4 naming would ensure that the files and
functions would still have unique names, but wouldn't need to be renamed
before adding the plugin to a site. Also, it allows for functions to be
reused from site to site without renaming.